Mar 30, 2024, 3:06 PM

+++ Item.cpp
@@ -30,7 +30,7 @@ int Item::getCostToShip() const
 {
   return costToShip;
 }
-double Item::getReward() const
+int Item::getReward() const
 {
   return reward;
 }
